+++ packages/svelte/src/index.ts
@@ -37,6 +37,7 @@ export { default as Timeline } from "./Timeline.svelte";
 export { default as Steps } from "./Steps.svelte";
 export { default as Rating } from "./Rating.svelte";
 export { default as Meter } from "./Meter.svelte";
+export { default as Empty } from "./Empty.svelte";
 export { default as Progress } from "./Progress.svelte";
 export { default as Slider } from "./Slider.svelte";
 export { default as ColorPicker } from "./ColorPicker.svelte";
@@ -107,6 +108,7 @@ export {
 	XtyleSteps,
 	XtyleRating,
 	XtyleMeter,
+	XtyleEmpty,
 	XtyleProgress,
 	XtyleSlider,
 	XtyleColorPicker,