+++ packages/xtyle/src/css/index.ts
@@ -12,6 +12,7 @@ export { timelineCss } from "./components/timeline.js";
 export { stepsCss } from "./components/steps.js";
 export { ratingCss } from "./components/rating.js";
 export { meterCss } from "./components/meter.js";
+export { emptyCss } from "./components/empty.js";
 export { parallaxCss } from "./components/parallax.js";
 export { heroCss } from "./components/hero.js";
 export { barCss } from "./components/bar.js";